+++ src/Service/Kms/src/Result/GetPublicKeyResponse.php
@@ -174,21 +174,13 @@ class GetPublicKeyResponse extends Result
         $data = $response->toArray();
 
         $this->keyId = isset($data['KeyId']) ? (string) $data['KeyId'] : null;
-        $this->publicKey = isset($data['PublicKey']) ? self::formatKey($data['PublicKey']) : null;
+        $this->publicKey = isset($data['PublicKey']) ? base64_decode((string) $data['PublicKey']) : null;
         $this->customerMasterKeySpec = isset($data['CustomerMasterKeySpec']) ? (string) $data['CustomerMasterKeySpec'] : null;
         $this->keySpec = isset($data['KeySpec']) ? (string) $data['KeySpec'] : null;
         $this->keyUsage = isset($data['KeyUsage']) ? (string) $data['KeyUsage'] : null;
         $this->encryptionAlgorithms = empty($data['EncryptionAlgorithms']) ? [] : $this->populateResultEncryptionAlgorithmSpecList($data['EncryptionAlgorithms']);
         $this->signingAlgorithms = empty($data['SigningAlgorithms']) ? [] : $this->populateResultSigningAlgorithmSpecList($data['SigningAlgorithms']);
         $this->keyAgreementAlgorithms = empty($data['KeyAgreementAlgorithms']) ? [] : $this->populateResultKeyAgreementAlgorithmSpecList($data['KeyAgreementAlgorithms']);
-    }
-
-    private static function formatKey(string $derData): string
-    {
-        $pem = "-----BEGIN PUBLIC KEY-----\n";
-        $pem .= chunk_split((string) $derData, 64, "\n");
-
-        return $pem . "-----END PUBLIC KEY-----\n";
     }
 
     /**