+++ src/Service/Kms/tests/Unit/Result/GetPublicKeyResponseTest.php
@@ -37,20 +37,7 @@ class GetPublicKeyResponseTest extends TestCase
         $result = new GetPublicKeyResponse(new Response($client->request('POST', 'http://localhost'), $client, new NullLogger()));
 
         self::assertSame('arn:aws:kms:us-west-2:111122223333:key/0987dcba-09fe-87dc-65ba-ab0987654321', $result->getKeyId());
-
-        $pubkey =
-            "-----BEGIN PUBLIC KEY-----\n" .
-            "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAyP/BlKX04RLmV8Q2VTSg\n" .
-            "TpJloxi6MBa1KQ3PMpelcRK4qbTX2+KLw1fiRWsoihvKyVDjvCoehSiv3gnHqzXo\n" .
-            "mMYk/9BxZlJV3R7bDCaeNbjgTem3Qgoxe/0MhYjGoGpNkryJnSJXy03Mx8WNA05Z\n" .
-            "YErvpt5YATaw7wQtJ+Mei6I/OxemaTQ6QqE5ulQ8wYeF96f4me6RYVuMHb5XcYkA\n" .
-            "1BMv+two3YCxvQ445EsE+HjExGvNL6Ot+g4yS+cwfcWIcU1kdVTTjXxQVqwQ3kRG\n" .
-            "b3G934XSyl5BeT7ItbTFNZGJVTNq3GAa3mAQv3BMXPsrhhqcZ/u+F+t2VKe93hw1\n" .
-            "+QIDAQAB\n" .
-            "-----END PUBLIC KEY-----\n";
-
-        self::assertSame($pubkey, $result->getPublicKey());
-
+        self::assertSame('MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAyP/BlKX04RLmV8Q2VTSgTpJloxi6MBa1KQ3PMpelcRK4qbTX2+KLw1fiRWsoihvKyVDjvCoehSiv3gnHqzXomMYk/9BxZlJV3R7bDCaeNbjgTem3Qgoxe/0MhYjGoGpNkryJnSJXy03Mx8WNA05ZYErvpt5YATaw7wQtJ+Mei6I/OxemaTQ6QqE5ulQ8wYeF96f4me6RYVuMHb5XcYkA1BMv+two3YCxvQ445EsE+HjExGvNL6Ot+g4yS+cwfcWIcU1kdVTTjXxQVqwQ3kRGb3G934XSyl5BeT7ItbTFNZGJVTNq3GAa3mAQv3BMXPsrhhqcZ/u+F+t2VKe93hw1+QIDAQAB', base64_encode($result->getPublicKey()));
         self::assertSame('RSA_4096', $result->getKeySpec());
         self::assertSame('ENCRYPT_DECRYPT', $result->getKeyUsage());
         self::assertContains('RSAES_OAEP_SHA_1', $result->getEncryptionAlgorithms());