+++ src/Service/CloudWatchLogs/src/CloudWatchLogsClient.php
@@ -333,6 +333,7 @@ class CloudWatchLogsClient extends AbstractApi
             case 'il-central-1':
             case 'me-central-1':
             case 'me-south-1':
+            case 'mx-central-1':
             case 'sa-east-1':
             case 'us-east-1':
             case 'us-east-2':