+++ src/Service/Lambda/src/LambdaClient.php
@@ -581,6 +581,7 @@ class LambdaClient extends AbstractApi
             case 'il-central-1':
             case 'me-central-1':
             case 'me-south-1':
+            case 'mx-central-1':
             case 'sa-east-1':
             case 'us-east-1':
             case 'us-east-2':