+++ src/Service/SecretsManager/src/SecretsManagerClient.php
@@ -518,6 +518,7 @@ class SecretsManagerClient extends AbstractApi
             case 'il-central-1':
             case 'me-central-1':
             case 'me-south-1':
+            case 'mx-central-1':
             case 'sa-east-1':
             case 'us-east-1':
             case 'us-east-1-fips':