+++ src/Service/Kms/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## NOT RELEASED
 
+- Added getPublicKey and verify operation
+
 ## 1.6.0
 
 ### Added