+++ src/Service/Kms/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Added getPublicKey and verify operation
-- AWS api-change: TODO
+- AWS api-change: Added `ap-southeast-7` and `ap-southeast-7-fips` regions
 
 ## 1.6.0